--- src/config.ts
@@ -1,4 +1,0 @@
-const PROD_HOST_URL: string = 'next://app';
-const RENDERER_OUTPUT_DIR: string = './src/renderer/out';
-
-export { PROD_HOST_URL, RENDERER_OUTPUT_DIR };
