--- src/config/index.ts
@@ -1,4 +1,0 @@
-const PROD_HOST_URL = 'https://web.ente.io';
-const FILE_STREAM_CHUNK_SIZE: number = 4 * 1024 * 1024;
-
-export { PROD_HOST_URL, FILE_STREAM_CHUNK_SIZE };
