+++ src/main/config/index.ts
@@ -1,0 +1,6 @@
+const PROD_HOST_URL: string = 'next://app';
+const RENDERER_OUTPUT_DIR: string = './src/renderer/out';
+
+const FILE_STREAM_CHUNK_SIZE: number = 4 * 1024 * 1024;
+
+export { PROD_HOST_URL, RENDERER_OUTPUT_DIR, FILE_STREAM_CHUNK_SIZE };
